pyproject.toml
[build-system]
requires = ["hatchling", "maturin>=1.0,<2.0"]
build-backend = "hatchling.build"

[project]
name = "nexus-fs"
version = "0.1.0"
description = "Ultra-Fast AI-Powered File Organization System for Windows - Faster than Everything Search"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "ShadowByte" }]
keywords = [
    "file-organization",
    "ai",
    "mcp",
    "semantic-search",
    "windows",
    "ultra-fast",
    "indexer"
]

dependencies = [
    # MCP Protocol
    "mcp>=1.0.0",

    # Core
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "typer[all]>=0.12.0",
    "rich>=13.0.0",
    "toml>=0.10.0",

    # AI/ML - Embeddings
    "sentence-transformers>=4.0.0",
    "torch>=2.0.0",
    "transformers>=4.35.0",

    # Vector Databases
    "qdrant-client>=1.16.0",
    "chromadb>=1.1.0",

    # Search
    "tantivy>=0.22.0",
    "whoosh>=2.7.0",  # Backup search engine

    # OCR
    "surya-ocr>=0.4.0;platform_system=='Windows'",
    "pytesseract>=0.3.10",
    "pdf2image>=1.16.0",

    # Document Processing
    "pypdf>=4.0.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.0",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12.0",

    # Image Processing
    "pillow>=10.0.0",
    "imagehash>=4.3.0",  # For duplicate detection

    # HTTP/Async
    "httpx>=0.27.0",
    "aiofiles>=24.0.0",
    "anyio>=4.0.0",

    # Utilities
    "watchdog>=4.0.0",
    "xxhash>=3.4.0",
    "psutil>=5.9.0",
    "humanize>=4.9.0",
    "python-dateutil>=2.8.0",
    "orjson>=3.9.0",  # Fast JSON
    "msgpack>=1.0.0",  # Fast serialization

    # Concurrency & Performance
    "uvloop>=0.19.0;platform_system!='Windows'",
    "multiprocess>=0.70.0",  # Better multiprocessing
    "joblib>=1.3.0",  # Parallel processing

    # Logging
    "structlog>=24.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "pytest-benchmark>=4.0.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

gui = [
    # Tauri dependencies managed by npm/cargo
]

full = [
    "nexus-fs[dev]",
    # Additional ML models
    "clip-interrogator>=0.6.0",
    "openai-whisper>=20231117",  # Audio transcription
]

[project.scripts]
nexus = "nexus_cli.main:app"
nexus-mcp = "nexus_mcp.__main__:main"
nexus-analyze = "nexus_ai.tools.space_analyzer:main"

[project.urls]
Repository = "https://github.com/yourusername/nexus-fs"

[tool.hatch.build.targets.wheel]
packages = ["src/nexus_ai", "src/nexus_mcp", "src/nexus_cli"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/configs",
    "/scripts",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.black]
line-length = 100
target-version = ["py311"]
